+++ infrastructure/Repositories/Auth/SocialRepository.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Infrastructure\Repositories;
+namespace Infrastructure\Repositories\Auth;
 
 use Domain\Entities\{
     RegisterUserEntity, RegisterUserNickNameEntity, SocialUserAccountEntity, RegisterSocialUserEntity
